% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/someRutils.R
\docType{package}
\name{someRutils}
\alias{someRutils}
\alias{_PACKAGE}
\alias{sumerutils}
\alias{someRutils-package}
\title{someRutils: A compilation of R data analysis utility functions}
\description{
A compilation of R utility functions for data cleaning, correction and visualisation
}
\author{
\strong{Maintainer}: Arnaud Wolfer \email{adwolfer@gmail.com}

}
